style(interfaces): add explicit public access modifiers to interface members
Standardize interface declarations by explicitly adding public access
modifiers to all interface methods and properties for code clarity
and consistency with project conventions.

- Add public modifier to interface methods across Core, World, and Entities projects
- Fix Engine.GetFramesPerSecond() return type from int to double in DebugOverlayNode
- Improve code readability and adherence to explicit declaration standards

+++ src/MineRPG.Godot.UI/DebugOverlayNode.cs
@@ -83,7 +83,7 @@ public sealed partial class DebugOverlayNode : Control
         float playerY = _debugData.PlayerY;
         float playerZ = _debugData.PlayerZ;
 
-        int framesPerSecond = Engine.GetFramesPerSecond();
+        double framesPerSecond = Engine.GetFramesPerSecond();
         double frameTimeMs = framesPerSecond > 0 ? MillisecondsPerSecond / framesPerSecond : 0;
         ulong memoryBytes = OS.GetStaticMemoryUsage();
         double memoryMegabytes = memoryBytes / BytesPerMegabyte;
